BAU — Stop Cypress tests breaking on New Year’s Day 🥳
The cy.clock(…) function only overrides what time the browser
thinks it is, not what the time is on the server.

This means that when running the Cypress tests, the example expiry
date shown on the enter card details page will always be two years
out from the actual current year. Which will mean it will change
from 10/22 to 10/23 on New Year’s Day, causing our Cypress test
that looks for 10/22 to fail.

Not a great start to the new year.

Fix this by making the test instead just look for 10/xx where xx
is any two digits.

We could try to calculate the current date at the time the test
runs but this would have unreliable results around midnight on New
Year’s Eve, when people are usually occupied by other things. We
have unit tests that check the year is two years out, so just
checking that there is an example expiry date in the right format
should be sufficient for the Cypress test.