PP-6328 Fix selecting Google Pay merchant ID by gateway account ID
The environment variable that says which gateway account IDs should
use GOOGLE_PAY_MERCHANT_ID_2 rather than GOOGLE_PAY_MERCHANT_ID is
a string containing space-separated IDs, which gets tokenised into
an array of strings. The gateway account ID we look for in the
array is a number, so it needs to be converted to a string first or
we’ll never find a match.