PP-66611 Handle Apple Pay not having shipping contact better
It is our understanding that if we receive an Apple Pay payload
with an email address, we get a shippingContact object with an
emailAddress property populated and other properties, such as
givenName and familyName, set to empty strings (because we
don’t actually request a shipping name when initiating Apple Pay).
When there is no email address (because email address collection
is switched off for the service), we believe we get no
shippingContact object at all (though we would if we requested a
shipping name).

Harden our code to remove the assumption that there’s always a
shippingContact object.