PP-7075 Add logging for DDC request for Worldpay 3DS flex
Add a hidden form field to the payment details form containing a status
for the Device Data Collection (DDC) for Worldpay 3DS flex.

Update this status at various points in performing the DDC request.
On the server side, log the value of this hidden status field so we can
use this to determine at what point the DDC went wrong if we do not get
a result to use.