PP-7351 Put HTTP Accept header in DOM of enter card details page
Make it so that the enter card details page’s HTML body element
has a data-accept-header attribute whose value is the contents of
the HTTP Accept request header.

This is so that it will be available to JavaScript when we’re doing
a Google Pay payment and can be read and sent off to the server in
an AJAX request (the AJAX request itself will likely have a
different Accept header, which is why we can’t use it).