BAU Bump JS Commons
This will pick up custom branding changes for MHRA and NIBSC reference
zen desk ticket https://govuk.zendesk.com/agent/tickets/4487202.

Note that release `3.1.0` was created in error about 10 days ago and was
subsequently corrected to `3.0.5` which I have confirmed with the
developer. It appears frontend wasn't updated with the corrected
version. This is not downgrading the version of `pay-js-commons`.