PP-8109 Remove writing the pid to file on startup
The container's entrypoint is `tini` which is a process manager and will
manage any orphaned processes. We want to run our applications within
containers with a readonly filesystem. We no longer need to write the
process id to a file to be read and used when the application recieves a
`SIGINT` because `tini` will take care of passing signals to the
processes beneath it.