PP-8858 Don't log error when apps return 500
500 indicates an internal error occurred. In this case, the app the
internal error occurred in should decide whether to log an error so that
it gets reported to Sentry. Currently, this causes a lot of noise in
Sentry.

Continue to log an error if the status code is >500 as this would
indicate that there has been some problem with communicating with the
app and likely needs investigating.