PP-10077 Make Google Pay and Apple Pay spinner cover whole page
Make it so that when we are processing a Google Pay or Apple Pay
payment we hide the majority of the page while showing the spinner
(currently none of the page is hidden). In addition, disable all
the submit buttons when the spinner is shown (currently only the
Google Pay or Apple Pay submit button is disabled).

This brings the appearance of the spinner into line with when we
do Worldpay 3DS Flex device data collection. Also make it so that
when the spinner is shown for DDC we disable all the submit buttons
(currently only the card details submit button is disabled).

with @SandorArpa