PP-10077 Nest Google Pay PaymentResponse object in Ajax request
When we make the Ajax request to
/web-payments-auth-request/google/CHARGE_ID, we currently
convert the PaymentResponse object received from Google Pay to
JSON and then add the two new worldpay3dsFlexDdcResult and
worldpay3dsFlexDdcStatus properties to it and then send that as
the payload.

Instead, send a JSON payload containing a top-level paymentResponse
property containing the unmodified PaymentResponse object from
Google Pay alongside the worldpay3dsFlexDdcResult and
worldpay3dsFlexDdcStatus properties, also at the top-level:

{
  "paymentResponse": … the PaymentResponse object from Google Pay …,
  "worldpay3dsFlexDdcResult": "the DDC result string",
  "worldpay3dsFlexDdcStatus": "the DDC status string"
}

This means we are not modifying the PaymentResponse object (which
is not really ours to modify).

Since Google Pay is currently disabled globally, it’s safe to
change both the code that creates the Ajax request and the server-
side code that processes the received Ajax request simultaneously
without having to worry about supporting both shapes of data
simultaneously for a while.

with @SandorArpa