PP-10077 Do not try DDC again if Google Pay auth fetch call fails
Remove the overcomplicated heroics that try to recreate the
Worldpay 3DS Flex device data collection iframe for Google Pay if
both device data collection and the Google Pay authorisation Ajax
fetch fail first time. On subsequent attempts, just try the
authorisation without a device data collection result.

It seems this never actually worked and it’s way too much code and
maintenance effort for an edge case where we’re already in a
failure mode.

We now store the device data collection status on the global Charge
object and use its presence to determine if we have already tried
to perform device data collection. We no longer check if the device
data collection iframe has been navigated away to another origin as
this should be now unnecessary. If the stored device data
collection status and the iframe origin get out of sync, the worst
that should happen is that device data collection will time out.

Add Cypress tests for this, including a fair amount of reworking
of existing tests to simplify them and remove hard-to-reproduce
failures.

with @SandorArpa