PP-10839: include google pay cross origin redirect in CSP
When a domain included in the content security policy results in a
cross-origin redirect, the redirected resource also needs to be included
in the policy to allow the connection.

We've seen Google Pay payments work since updating the CSP to allow
`google.com/pay` however we're continuing to see CSP violations. If
we're continuing to block connects we may see Google Pay payment
continue to silently fail, include the redirect address to allow
connection to both.