BAU: Remove Docker dependabot version restrictions
We are now firmly on node16 so the old 'ignore' config is no longer needed. This old config also might be the reason Dependabot is not providing updates for Node base images - the pay-stream-s3-sqs repo has the same Dockerfile setup but no Dependabot restrictions, and the updates are appearing as expected (see https://github.com/alphagov/pay-stream-s3-sqs/pull/579 for example).