PP-11082: update google pay csp to include www subdomain
We're seeing CSP rule violations for Google Pay, starting on Friday 26th
May and ramping up Tuesday 30th May. Investigating this it looks like
Google is now trying to connect to the `www.` subdomain rather than the
previously used `google.com`. This observation doesn't come from an
official source so we should see reason about if it resolves the issue
before locking in the change.