PP-11244 Prepare for changing find charge by token response
Want to change the response schema for the connector GET
`/v1/frontend/tokens/{chargeTokenId}` endpoint to use the
FrontendChargeResponse schema for the embedded charge object in the
response, rather than ChargeEntity.

Support both versions of the response temporarily by checking which
fields are returned in the response. This will be updated again to
remove support for the old response schema after the connector change
is made.

Alter the Pact test for this endpoint to not expect the response to
contain the `charge.externalId` field to allow the connector to be
updated.