PP-11244 Remove handling for old GET token response
Connector has now been updated so that the GET token response uses
the FrontendChargeResponse schema for the embedded charge rather than
ChargeEntity.

Remove the handling for the old response.

Update `secure.controller.test.js` for this change, and refactor it
slightly to make it easier to understand.

Update the Pact test for the GET token endpoint to validate the
embedded charge in the response using the new schema.