PP-11323 hide wallet options for recurring payments
We currently only offer wallet payments for one-off payments. Although it is possible to use recurring payments (payments taken when setting up agreements), it needs additional backend integration which is not in place for both Stripe/Worldpay.

So we should hide wallet options even if allow_apple_pay / allow_google_pay is true for the gateway account

- do not render wallet payments if the payment is a recurring one (savePaymentInstrumentToAgreement is present)
- add Cypress tests