PP-11363 Use correct Apple Pay certificate for payment provider
Determine which certificate environment variables to use
(stripe/worldpay) by looking at the payment provider for the charge.

Convert tests to unit tests using sinon/proxyquire to make them easier
to write/debug.

Add in additional logging for the expiry date for the Stripe certificate
making sure not to break the Splunk alert for when the certificate is
close to expiring. The Splunk alert will fire for both Worldpay/Stripe