PP-11462 fix erroneous apple pay option (#3680)
In some circumstances, when a user encountered an error during the Google Pay wallet journey they were returned to the card details page and could see Apple Pay as a payment option.

This was because a client-side script modified the DOM to show/hide wallet options. By wrapping all options in a parent div and applying conditional styling higher up the document tree, we no longer accidentally toggle the visibility of Apple Pay-specific DOM elements.