PP-11638 Send client-side errors to Sentry
A proof of concept for sending client-side errors to Sentry. This will
send all unhandled errors from our client.

Things to consider before we turn this on:

- What are the privacy considerations? Could there be sensistive data
in JS exceptions? Can we prevent against this in a reliable way?
- Do we need to update our privacy notice if we are including Sentry
JS?
- Will we run into high costs from enabling this? Are there lots of
exceptions that will be logged, or is there a potential for abuse?