PP-11638 Add endpoint for client-side logging
Add a `/log/:chargeId` POST route. This can be called by the
client-side JavaScript to do logging.

This accepts a POST request with a body containing a `code`, which
maps to a message to be logged. We want to do this rather than
accepting any string to be logged to avoid log injection.

This endpoint always returns a 200, even if we do not log due to the
`code` being invalid or missing. This is because we do not want to
interrupt client-side JavaScript running if this fails.

In the client-side JavaScript call this endpoint to log steps in the
wallet payment journeys, such as when the button is clicked to start
an Apple Pay or Google Pay payment, and also to log when certain
errors are caught. When we call the loggng endpoint, we do not
handle the Promise resolving/rejecting in most cases, as we don't need
to handle the response in any way. Also for initiating an Apple Pay
payment, for example, chaining promises causes the request to start
the Apple Pay payment to fail.