PP-11638 Improve logging for handling wallet auth response (#3764)
Log the "wallet" on log lines if it is available on the charge. This
will only be present post-authorisation.

Log a line when we are handling the authorisation response from
connector. This is so we can easily validate that the redirect occurs
in the client-side JavaScript after the authorisation is completed.

Stop trying to send a log to the server from the client-side
JavaScript when the authorisation has been processed. There is no
garuantee this will complete before the page redirects meaning we were
dropping some log lines. We could move the redirect to a `then` clause
to execute after the promise to send the log message has resolved, but
this would slow down the page redirect, which is undesirable.

Tidy up some of the client-side JavaScript slightly. The `init`
function in `web-payments/index.js` no longer takes any parameters.