PP-11638 Add wallet path param for logging (#3768)
Relying on getting the wallet_type from the charge when handling the
authorisation result isn't reliable - as connector can return a
response before it has completed authorisation, in which case the
wallet_type is not present.

So stop logging the `wallet` from the `wallet_type` on the charge from
connector, as we can't rely on it consistently being there.

Instead, add the wallet type as a path parameter for the
/handle-payment-response/:chargeId path, and pass in this wallet type
when passing the redirect URL back to the client-side JS in
`payment-auth-request.controller.js`. We will use this for logging in
`handle-auth-response.controller.js`.

For now, also support the old URL without the wallet as a path
parameter so we don't break in-flight payments. Support for this will
be removed as a follow-up commit. We will also the logging of the
wallet type in this follow-up commit. We can't add this yet, as the
path parameter will not be present for in-flight payments when this
change is deployed.

Rename some path params from `provider` -> `wallet` to avoid confusion
with payment provider.