PP-11638 Log wallet type when handling auth response
c4ca3a43ec61b017216baa5dc67088e992927c89 added the wallet type as a
path parameter to the `/handle-payment-response/:wallet/:chargeId`
route. Use this in log lines to log a `wallet` field, which is either
"apple" or "google".

Remove the old `/handle-payment-response/:chargeId` route, which was
maintained so as not to break in-flight payments when the new route
with the path parameter was introduced. New payment journeys will
already be directed to the route with the path parameter.