PP-12853 Update apple pay merchant validation to use Axios
With this change, we are removing the use of `requestretry` from the Merchant Validation Controller which currently uses this module.

We are replacing it by using Axios directly.

There is an [Axios bug](https://github.com/axios/axios/issues/4531) when
using an SSL proxy.

Therefore, we have had to add the NPM [https-proxy-agent](https://www.npmjs.com/package/https-proxy-agent) as a workaround.

Note that the previous attempted did not pass our QA on the test environment[1].

Further information in Jira[2].

[1]
https://github.com/alphagov/pay-frontend/pull/3862

[2]
https://payments-platform.atlassian.net/browse/PP-12853