PP-12161 Enable test for disabled Apple Pay option
With this change, we are enabling the test to check for when Apple Pay is not
supported by the browser.

The test was previously commented out as there was a problem with testing that
the page was still showing the payment details.

Now that we are checking directly and explicitly for the presence of the
payment anount, the test passes and can be re-enabled.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-12161