PP-12161 Prevent duplication of IDs within payment summary
With this change, we are updating the payment summary macro so that when it is
called for a second time, considering different scenarios, it sets its
elements with different IDs.

This allows us also to enable a test to check when Apple Pay is not supported
by the browser, which was previously created in a separate branch and
commented out as it would not work given the repetition of IDs caused by the
payment summary macro.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-12161