Use axios for ApplePay Merchant Validation with or without proxy
With this change, we are updating the controller in charge of operating the
Apple Pay Merchant Validation so that it uses axios instead of request retry,
and also so that it uses HttpsProxyAgent if a proxy URL is defined in the
environment.

This is needed because we have realised that the test environment uses a
proxy, while the local environment does not, thus requiring us to use axios
in two different ways depending on the presence or absence of a proxy.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-12853