Update Apple Pay Merchant Validation with Axios
With this change, we are updating the implementation of the Apple Pay Merchant
Validation so that when Axios is enabled (which is the case on Test only, for
now), then we pass the HttpsAgent or the HttpsProxyAgent to Axios correctly
based on the presence or absence of a proxy URL.

We are also adding a log to check that the proxy URL is correctly setup,
although we are not logging the URL itself for security reasons.

Further information in Jira.

https://payments-platform.atlassian.net/browse/PP-12853