PP-12853 Add the charset in Content-Type header for Apple Pay
With the previous PR[1], we have made some progress in communicating with
Apple Pay via Axios and HttpsProxyAgent and we have tested that the
validation is initiated correctly but then fails with the following message
from Apple: `Apple Pay Merchant ID not valid`.

See screenshot in the new PR[2].

With this change, we are now adding the charset to the Content-Type header in
the post request, in case that is what is expected by Apple Pay, and we are
also adding the status to the error log, to have a little bit more
information back.

Further information in JIRA[3].

[1]
https://github.com/alphagov/pay-frontend/pull/3885

[2]
https://github.com/alphagov/pay-frontend/pull/3886

[3]
https://payments-platform.atlassian.net/browse/PP-12853