PP-12853 Pass ProxyAgent to HttpsAgent for Apple Pay validation with axios
With the previous PRs[1][2][3], we have made some progress in communicating
with Apple Pay via Axios and HttpsProxyAgent and we have tested that the
validation is initiated correctly but then fails with the following message:
`Apple Pay Merchant ID not valid`.

See screenshot in the second PR[2].

With this change, we are now creating the HttpsProxyAgent with just the proxy
URL as an argument for its constructor, passing it together with the
certificate and key to the HttpsAgent, as arguments for its constructor, and
then creating an `axios` instance, passing the HttpsAgent to it.

Further information in JIRA[3].

[1]
https://github.com/alphagov/pay-frontend/pull/3885

[2]
https://github.com/alphagov/pay-frontend/pull/3886

[3]
https://github.com/alphagov/pay-frontend/pull/3887

[4]
https://payments-platform.atlassian.net/browse/PP-12853