Update path in static files for Apple Pay association file
With this change, we are updating the path where the Apple Pay association
file needs to be placed in order for the pay-frontend application to serve it
as a static file so that Apple can verify our domain.

This change will be made in the instructions of the Pay Team manual too, at
the step no. 8 of the section `Testing Apple Pay locally for Stripe`, linked
below.

https://manual.payments.service.gov.uk/manual/how-to/test-apple-and-google-pay.html#testing-apple-pay-locally-for-stripe