Remove code path for Apple Pay validation with request retry
With a recent change[1], we have enabled Apple Pay Merchant Validaton with
`axios` and `hpagent`.

The change has been successfully verified on the Test environment, where the
code path with `axios` was switched on using an environment variable[2].

Now that the code path with `axios` has been switched ON also on Production,
then we are now able to stop using `requestretry` in pay-frontend by
completely deleting the code path using it for the Apple Pay Merchant
Validation.

Note that whilst working on this code, we have also improved the Pay CLI, and
that now it is possible to run Pay locally --with-egress-proxy. For this
reason, we are also removing the code path using `axios` without a proxy URL,
since it is not production like and it is not needed for local development
either.

Further information in JIRA[3].

[1]
https://github.com/alphagov/pay-frontend/pull/3891

[2]
https://github.com/alphagov/pay-infra/pull/4913

[3]
https://payments-platform.atlassian.net/browse/PP-12853