Remove code path for Apple Pay validation with request retry
With a recent change[1], we have enabled Apple Pay Merchant Validaton with
`axios` and `hpagent`.

The change has been successfully verified on the Test environment, where the
code path with `axios` was switched on using an environment variable[2].

Once the code path with `axios` has been switched off also on Production, then
we will be able to stop using `requestretry` in pay-frontend by completely
deleting the code path using it for the Apple Pay Merchant Validation.

Further information in JIRA[3].

[1]
https://github.com/alphagov/pay-frontend/pull/3891

[2]
https://github.com/alphagov/pay-infra/pull/4913

[3]
https://payments-platform.atlassian.net/browse/PP-12853